+++ Riot/Modules/Application/LegacyAppDelegate.m
@@ -1866,6 +1866,8 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
             [self registerNewRequestNotificationForSession:mxSession];
             
             [self checkLocalPrivateKeysInSession:mxSession];
+            
+            [self.pushNotificationService checkPushKitPushersInSession:mxSession];
         }
         else if (mxSession.state == MXSessionStateClosed)
         {