+++ Tools/Release/buildRelease.sh
@@ -38,24 +38,16 @@ bundle update
 mkdir -p $BUILD_DIR
 cd $BUILD_DIR
 REPO_URL=$(git ls-remote --get-url origin)
-git clone $REPO_URL
 REPO_NAME=$(basename -s .git $REPO_URL)
+git clone $REPO_URL --depth=1 --branch $TAG
 cd $REPO_NAME
-git checkout $TAG $TAG
 
 
-# Develop branch special case
-if [ "$TAG" = "develop" ]; then 
-    ./use-dev-pods.sh
-fi
-
+# Use appropriated dependencies according to the current branch
+bundle exec fastlane point_dependencies_to_same_feature
 
 # Build
 bundle exec fastlane app_store build_number:$BUILD_NUMBER
-
-# Check ipa
-./checkipa.sh out/Riot.ipa
-
 
 if [ -e out/Riot.ipa ]; then
     # Here is the artefact